+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/assets/asset_base.py
@@ -108,7 +108,7 @@ class AssetBase(ABC):
     def has_debug_vis_implementation(self) -> bool:
         """Whether the asset has a debug visualization implemented."""
         # check if function raises NotImplementedError
-        source_code = inspect.getsource(self._debug_vis_callback)
+        source_code = inspect.getsource(self._set_debug_vis_impl)
         return "NotImplementedError" not in source_code
 
     """